+++ tests/CMakeLists.txt
@@ -9,7 +9,7 @@ set(THREADS_PREFER_PTHREAD_FLAG ON)
 
 find_package(Threads REQUIRED)
 
-add_executable (tests main.cpp communication_tests.cpp time_out_tests.cpp pseudoterminal.cpp helpers.cpp)
+add_executable (tests main.cpp communication_tests.cpp time_out_tests.cpp pseudoterminal.cpp common.cpp)
 
 target_link_libraries(tests PRIVATE lin_serial Threads::Threads)
