+++ tests/common.cpp
@@ -2,7 +2,7 @@
 #include <memory>
 #include <vector>
 #include <iostream>
-#include "helpers.h"
+#include "common.h"
 
 std::string execute(const std::string& cmd)
 {
