+++ tests/common.h
